+++ Example3.InvoiceCode/EfCoreCode/InvoiceTenantChangeService.cs
@@ -12,6 +12,9 @@ namespace Example3.InvoiceCode.EfCoreCode
 {
     public class InvoiceTenantChangeService : ITenantChangeService
     {
+        /// <summary>
+        /// This creates an instance of the application's DbContext to use within an transaction with the AuthPermissionsDbContext
+        /// </summary>
         public DbContext GetNewInstanceOfAppContext(SqlConnection sqlConnection)
         {
             var options = new DbContextOptionsBuilder<InvoicesDbContext>()
@@ -22,6 +25,15 @@ namespace Example3.InvoiceCode.EfCoreCode
             return new InvoicesDbContext(options, null);
         }
 
+        /// <summary>
+        /// When a new AuthP Tenant is created, then this method is called. If you have a tenant-type entity in your
+        /// application's database, then this allows you to create a new entity for the new tenant
+        /// </summary>
+        /// <param name="appTransactionContext">The application's DbContext within a transaction</param>
+        /// <param name="dataKey">The DataKey of the tenant being deleted</param>
+        /// <param name="tenantId">The TenantId of the tenant being deleted</param>
+        /// <param name="fullTenantName">The full name of the tenant being deleted</param>
+        /// <returns>Returns null if all OK, otherwise the delete is rolled back and the return string is shown to the user</returns>
         public async Task<string> CreateNewTenantAsync(DbContext appTransactionContext, string dataKey, int tenantId, string fullTenantName)
         {
             var newCompanyTenant = new CompanyTenant
@@ -36,6 +48,16 @@ namespace Example3.InvoiceCode.EfCoreCode
             return null;
         }
 
+        /// <summary>
+        /// This is called within a transaction to allow the the application-side of the database to either
+        /// a) delete all the application-side data with the given DataKey, or b) list the changes to show to the admin user
+        /// Note: The created application's DbContext won't have a DataKey, so you will need to use IgnoreQueryFilters on any EF Core read
+        /// </summary>
+        /// <param name="appTransactionContext">The application's DbContext within a transaction</param>
+        /// <param name="dataKey">The DataKey of the tenant being deleted</param>
+        /// <param name="tenantId">The TenantId of the tenant being deleted</param>
+        /// <param name="fullTenantName">The full name of the tenant being deleted</param>
+        /// <returns>Returns null if all OK, otherwise the delete is rolled back and the return string is shown to the user</returns>
         public async Task<string> HandleTenantDeleteAsync(DbContext appTransactionContext, string dataKey, int tenantId, string fullTenantName)
         {
             var deleteSalesSql = $"DELETE FROM invoice.{nameof(InvoicesDbContext.LineItems)} WHERE DataKey = '{dataKey}'";
@@ -44,6 +66,7 @@ namespace Example3.InvoiceCode.EfCoreCode
             await appTransactionContext.Database.ExecuteSqlRawAsync(deleteStockSql);
 
             var companyTenant = await appTransactionContext.Set<CompanyTenant>()
+                .IgnoreQueryFilters()
                 .SingleOrDefaultAsync(x => x.AuthPTenantId == tenantId);
             if (companyTenant != null)
             {
@@ -54,9 +77,19 @@ namespace Example3.InvoiceCode.EfCoreCode
             return null;
         }
 
+        /// <summary>
+        /// This is called when the name of your Tenants is changed. This is useful if you use the tenant name in your multi-tenant data.
+        /// NOTE: The created application's DbContext won't have a DataKey, so you will need to use IgnoreQueryFilters on any EF Core read
+        /// </summary>
+        /// <param name="appTransactionContext">The application's DbContext within a transaction</param>
+        /// <param name="dataKey">The DataKey of the tenant</param>
+        /// <param name="tenantId">The TenantId of the tenant</param>
+        /// <param name="fullTenantName">The full name of the tenant</param>
+        /// <returns>Returns null if all OK, otherwise the delete is rolled back and the return string is shown to the user</returns>
         public async Task<string> HandleUpdateNameAsync(DbContext appTransactionContext, string dataKey, int tenantId, string fullTenantName)
         {
             var companyTenant = await appTransactionContext.Set<CompanyTenant>()
+                .IgnoreQueryFilters()
                 .SingleOrDefaultAsync(x => x.AuthPTenantId == tenantId);
             if (companyTenant != null)
             {
@@ -67,6 +100,20 @@ namespace Example3.InvoiceCode.EfCoreCode
             return null;
         }
 
+        /// <summary>
+        /// This is used with hierarchical tenants, where you move one tenant (and its children) to another tenant
+        /// This requires you to change the DataKeys of each application's tenant data, so they link to the new tenant.
+        /// Also, if you contain the name of the tenant in your data, then you need to update its new FullName
+        /// Notes:
+        /// - The created application's DbContext won't have a DataKey, so you will need to use IgnoreQueryFilters on any EF Core read
+        /// - You can get multiple calls if move a higher level
+        /// </summary>
+        /// <param name="appTransactionContext"></param>
+        /// <param name="oldDataKey">The old DataKey to look for</param>
+        /// <param name="newDataKey">The new DataKey to change to</param>
+        /// <param name="tenantId">The TenantId of the tenant being moved</param>
+        /// <param name="newFullTenantName">The new full name of the tenant</param>
+        /// <returns>Returns null if all OK, otherwise the move is rolled back and the return string is shown to the user</returns>
         public Task<string> MoveHierarchicalTenantDataAsync(DbContext appTransactionContext, string oldDataKey, string newDataKey, int tenantId,
             string newFullTenantName)
         {